+++ cc-controlplane/outputs.tf
@@ -8,10 +8,12 @@ output "resource-ids" {
 }
 
 output "vpc_endpoint_service_name" {
-  value = confluent_private_link_attachment.pla.aws[0].vpc_endpoint_service_name
+  value = var.privatelink_mode == "platt" ? confluent_private_link_attachment.pla[0].aws[0].vpc_endpoint_service_name : confluent_gateway.main[0].aws_ingress_private_link_gateway[0].vpc_endpoint_service_name
 }
+
 output "dns_domain" {
-  value = confluent_private_link_attachment.pla.dns_domain
+  value = var.privatelink_mode == "platt" ? confluent_private_link_attachment.pla[0].dns_domain : ""
+  description = "DNS domain for PLATT mode. For gateway mode, use confluent_access_point.gateway output from main.tf"
 }
 
 output "app-manager" {
@@ -27,8 +29,13 @@ output "app-producer" {
 output "enterprise_cluster" {
   value = confluent_kafka_cluster.enterprise
 }
-output "plac" {
-  value = confluent_private_link_attachment_connection.plac
+
+output "pla" {
+  value = var.privatelink_mode == "platt" ? confluent_private_link_attachment.pla[0] : null
+}
+
+output "gateway" {
+  value = var.privatelink_mode == "gateway" ? confluent_gateway.main[0] : null
 }
 
 output "am-cluster-admin" {